+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.4.4] - 2023-09-04
+
 - Support EnvArg for whenEnvVar
 - Add help methods to build EnvArg
 - For top level condition of stage or pipeline it should combine all condition with && rule